Renamed manager/keyboard module to keyboard_handler
