+++ src/manager/mod.rs
@@ -1,11 +1,11 @@
 mod input_manager;
 mod output_manager;
-mod keyboard;
+mod keyboard_handler;
 mod pointer;
 mod output;
 
 pub use self::input_manager::{InputManager, InputManagerHandler};
-pub use self::keyboard::{KeyboardHandler, KeyboardWrapper};
+pub use self::keyboard_handler::{KeyboardHandler, KeyboardWrapper};
 pub use self::output::{Output, OutputHandler};
 pub use self::output_manager::{OutputManager, OutputManagerHandler};
 pub use self::pointer::{Pointer, PointerHandler};
